+++ src/styles/animation/index.js
@@ -1,0 +1,1 @@
+export * from './animation.styles';
